Refactor FROST protocol into protocol crates
