+++ crates/protocols/frost/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "mpc-protocols-frost"
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
+anyhow = { workspace = true }
+async-trait = { workspace = true }
+mpc-network = { workspace = true }
+mpc-protocols-codecs = { workspace = true }
+serde = { workspace = true }
+
+frost-ed25519 = { workspace = true }
+frost-secp256k1 = { workspace = true }
+frost-core = { workspace = true }
+rand = { workspace = true }
