+++ crates/protocols/frost/src/lib.rs
@@ -1,0 +1,9 @@
+mod keygen;
+mod keysign;
+mod messages;
+mod types;
+
+pub use keygen::run_dkg;
+pub use keysign::run_signing;
+pub use messages::{IncomingMessage, MessageReceiver, MessageSender, OutgoingMessage};
+pub use types::{KeyShare, PublicKey, Signature};
