+++ crates/tss/Cargo.toml
@@ -7,19 +7,16 @@ edition.workspace = true
 # Internal workspace crates
 mpc-network = { workspace = true }
 mpc-runtime = { workspace = true }
+mpc-protocols-codecs = { workspace = true }
+mpc-protocols-frost = { workspace = true }
 
 # Async runtime
 futures = { workspace = true }
-futures-util = { workspace = true }
 async-channel = { workspace = true }
-tokio = { workspace = true }
 async-trait = { workspace = true }
 
 # Error handling
 anyhow = { workspace = true }
-
-# Logging
-
 
 # Networking
 libp2p = { workspace = true }
@@ -28,17 +25,7 @@ unsigned-varint = { workspace = true }
 # Serialization
 serde = { workspace = true }
 serde_json = { workspace = true }
-serde_ipld_dagcbor = { workspace = true }
 
 # FROST
 frost-ed25519 = { workspace = true }
 frost-secp256k1 = { workspace = true }
-frost-core = { workspace = true }
-rand = { workspace = true }
-
-# Cryptography
-blake2 = { workspace = true }
-
-# Utilities
-strum = { workspace = true }
-hex = { workspace = true }
