+++ crates/tss/src/lib.rs
@@ -1,5 +1,6 @@
 extern crate core;
 
+mod adapter;
 mod config;
 mod factory;
 mod keygen;
@@ -8,4 +9,5 @@ mod keysign;
 pub use config::*;
 pub use factory::*;
 pub use keygen::*;
-pub use keysign::{PublicKey, Signature};
+pub use keysign::*;
+pub use mpc_protocols_frost::{PublicKey, Signature};
